Remove bevy_infite_grid
The bevy_infinite_grid dependency looks like it accomplishes what we want, but it has some weird bugs (e.g. not actually an infinite grid) and isn't very configurable.  We're probably going to have to fork it to make significant changes.  As it is the last dependency which does not support Bevy 0.18, we'll drop it for now and consider writing our own.