Persist application settings across runs using sqlite3 configuration database (#46)
Like the title says, this PR adds support for persistent settings using a sqlite3 database. There are a number of ways this implementation can probably be improved, such as moving the code into its own Bevy plugin module, reducing some boilerplate code duplication, and creating a consistent internal interface for persisting settings (instead of a one-off thing just for the primary window config).

But perfect is the enemy of good enough, and this PR improves the user experience.

* Add AppConfig struct which contains persistent settings about the primary Window.

* Persist application settings across runs using a sqlite configuration database.

* Use bundled sqlite3 instead of platform-provided.

* Restrict sqlite3 code to only run on desktop platforms.

A different solution will have to be investigated for iOS and Android.